[Refactor] 각 메소드를 적절한 위치로 변경
